refactor(util): InputParser를 util 클래스로 변경하고 util 클래스들을 util 패키지로 이동
