+++ src/main/java/util/AttendanceDateTimeFormatter.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 import java.time.format.DateTimeFormatter;
