+++ src/test/java/util/InputParserTest.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
